+++ pkg/gtk/v3/gtkrange.go
@@ -106,8 +106,8 @@ func init() {
 		GType:        GTypeRange,
 		GoType:       reflect.TypeOf((*Range)(nil)),
 		InitClass:    initClassRange,
-		ClassSize:    uint16(unsafe.Sizeof(C.GtkRange{})),
-		InstanceSize: uint16(unsafe.Sizeof(C.GtkRangeClass{})),
+		ClassSize:    uint32(unsafe.Sizeof(C.GtkRange{})),
+		InstanceSize: uint32(unsafe.Sizeof(C.GtkRangeClass{})),
 	})
 }
